♻️ (env): Move `top_k_metrics_and_plots` to base env
