+++ gflownet/envs/base.py
@@ -5,6 +5,8 @@ from abc import abstractmethod
 from copy import deepcopy
 from typing import List, Optional, Tuple, Union
 
+import matplotlib as mpl
+import matplotlib.pyplot as plt
 import numpy as np
 import numpy.typing as npt
 import torch
@@ -12,6 +14,8 @@ from torch.distributions import Categorical
 from torchtyping import TensorType
 
 from gflownet.utils.common import set_device, set_float_precision
+
+CMAP = mpl.colormaps["cividis"]
 
 
 class GFlowNetEnv:
@@ -589,3 +593,297 @@ class GFlowNetEnv:
         proxy_reward = self.proxy2reward(proxy_energy)
 
         return gt_energy, proxy_energy, gt_reward, proxy_reward
+
+    @torch.no_grad()
+    def top_k_metrics_and_plots(
+        self,
+        states,
+        top_k,
+        name,
+        energy=None,
+        reward=None,
+        step=None,
+        **kwargs,
+    ):
+        """
+        Compute top_k metrics and plots for the given states.
+
+        In particular, if no states, energy, or reward are passed, then the name
+        *must* be "train", and the energy and reward will be computed from the
+        proxy using `env.compute_train_energy_proxy_and_rewards()`. In this case,
+        `top_k_metrics_and_plots` will be called a second time to compute the
+        metrics and plots of the proxy distribution in addition to the ground-truth
+        distribution.
+        Train mode should only be called once at the begining of training as
+        distributions do not change over time.
+
+        If `states` are passed, then the energy and reward will be computed from the
+        proxy for those states. They are typically sampled from the current GFN.
+
+        Otherwise, energy and reward should be passed directly.
+
+        *Plots and metrics*:
+        - mean+std of energy and reward
+        - mean+std of top_k energy and reward
+        - histogram of energy and reward
+        - histogram of top_k energy and reward
+
+
+        Args
+        ----
+        states: list
+            List of states to compute metrics and plots for.
+
+        top_k: int
+            Number of top k states to compute metrics and plots for.
+            "top" means lowest energy/highest reward.
+
+        name: str
+            Name of the distribution to compute metrics and plots for.
+            Typically "gflownet", "random" or "train". Will be used in
+            metrics names like `f"Mean {name} energy"`.
+
+        energy: torch.Tensor, optional
+            Batch of pre-computed energies
+
+        reward: torch.Tensor, optional
+            Batch of pre-computed rewards
+
+        step: int, optional
+            Step number to use for the plot title.
+
+        Returns
+        -------
+        metrics: dict
+            Dictionary of metrics: str->float
+
+        figs: list
+            List of matplotlib figures
+
+        figs_names: list
+            List of figure names for `figs`
+        """
+
+        if states is None and energy is None and reward is None:
+            assert name == "train"
+            (
+                energy,
+                proxy,
+                energy_reward,
+                proxy_reward,
+            ) = self.compute_train_energy_proxy_and_rewards()
+            name = "train ground truth"
+            reward = energy_reward
+        elif energy is None and reward is None:
+            x = torch.stack([self.state2proxy(s) for s in states])
+            energy = self.proxy(x.to(self.device)).cpu()
+            reward = self.proxy2reward(energy)
+
+        assert energy is not None and reward is not None
+
+        # select top k best energies and rewards
+        top_k_e = torch.topk(energy, top_k, largest=False, dim=0).values.numpy()
+        top_k_r = torch.topk(reward, top_k, largest=True, dim=0).values.numpy()
+
+        # find best energy and reward
+        best_e = torch.min(energy).item()
+        best_r = torch.max(reward).item()
+
+        # to numpy to plot
+        energy = energy.numpy()
+        reward = reward.numpy()
+
+        # compute stats
+        mean_e = np.mean(energy)
+        mean_r = np.mean(reward)
+
+        std_e = np.std(energy)
+        std_r = np.std(reward)
+
+        mean_top_k_e = np.mean(top_k_e)
+        mean_top_k_r = np.mean(top_k_r)
+
+        std_top_k_e = np.std(top_k_e)
+        std_top_k_r = np.std(top_k_r)
+
+        # automatic color scale
+        # currently: cividis colour map
+        colors = ["full", "top_k"]
+        normalizer = mpl.colors.Normalize(vmin=0, vmax=len(colors) - 0.5)
+        colors = {k: CMAP(normalizer(i)) for i, k in enumerate(colors[::-1])}
+
+        # two sublopts: left is energy, right is reward
+        fig, ax = plt.subplots(1, 2, figsize=(10, 5))
+
+        # energy full distribution and stats lines
+        ax[0].hist(
+            energy,
+            bins=100,
+            alpha=0.35,
+            label=f"All = {len(energy)}",
+            color=colors["full"],
+            density=True,
+        )
+        ax[0].axvline(
+            mean_e,
+            color=colors["full"],
+            linestyle=(0, (5, 10)),
+            label=f"Mean = {mean_e:.3f}",
+        )
+        ax[0].axvline(
+            mean_e + std_e,
+            color=colors["full"],
+            linestyle=(0, (1, 10)),
+            label=f"Std = {std_e:.3f}",
+        )
+        ax[0].axvline(
+            mean_e - std_e,
+            color=colors["full"],
+            linestyle=(0, (1, 10)),
+        )
+
+        # energy top k distribution and stats lines
+        ax[0].hist(
+            top_k_e,
+            bins=100,
+            alpha=0.7,
+            label=f"Top k = {top_k}",
+            color=colors["top_k"],
+            density=True,
+        )
+        ax[0].axvline(
+            mean_top_k_e,
+            color=colors["top_k"],
+            linestyle=(0, (5, 10)),
+            label=f"Mean = {mean_top_k_e:.3f}",
+        )
+        ax[0].axvline(
+            mean_top_k_e + std_top_k_e,
+            color=colors["top_k"],
+            linestyle=(0, (1, 10)),
+            label=f"Std = {std_top_k_e:.3f}",
+        )
+        ax[0].axvline(
+            mean_top_k_e - std_top_k_e,
+            color=colors["top_k"],
+            linestyle=(0, (1, 10)),
+        )
+        # energy title & legend
+        ax[0].set_title(
+            f"Energy distribution for {top_k} vs {len(energy)}"
+            + f" samples\nBest: {best_e:.3f}",
+            y=0,
+            pad=-20,
+            verticalalignment="top",
+            size=12,
+        )
+        ax[0].legend()
+
+        # reward full distribution and stats lines
+        ax[1].hist(
+            reward,
+            bins=100,
+            alpha=0.35,
+            label=f"All = {len(reward)}",
+            color=colors["full"],
+            density=True,
+        )
+        ax[1].axvline(
+            mean_r,
+            color=colors["full"],
+            linestyle=(0, (5, 10)),
+            label=f"Mean = {mean_r:.3f}",
+        )
+        ax[1].axvline(
+            mean_r + std_r,
+            color=colors["full"],
+            linestyle=(0, (1, 10)),
+            label=f"Std = {std_r:.3f}",
+        )
+        ax[1].axvline(
+            mean_r - std_r,
+            color=colors["full"],
+            linestyle=(0, (1, 10)),
+        )
+
+        # reward top k distribution and stats lines
+        ax[1].hist(
+            top_k_r,
+            bins=100,
+            alpha=0.7,
+            label=f"Top k = {top_k}",
+            color=colors["top_k"],
+            density=True,
+        )
+        ax[1].axvline(
+            mean_top_k_r,
+            color=colors["top_k"],
+            linestyle=(0, (5, 10)),
+            label=f"Mean = {mean_top_k_r:.3f}",
+        )
+        ax[1].axvline(
+            mean_top_k_r + std_top_k_r,
+            color=colors["top_k"],
+            linestyle=(0, (1, 10)),
+            label=f"Std = {std_top_k_r:.3f}",
+        )
+        ax[1].axvline(
+            mean_top_k_r - std_top_k_r,
+            color=colors["top_k"],
+            linestyle=(0, (1, 10)),
+        )
+
+        # reward title & legend
+        ax[1].set_title(
+            f"Reward distribution for {top_k} vs {len(reward)}"
+            + f" samples\nBest: {best_r:.3f}",
+            y=0,
+            pad=-20,
+            verticalalignment="top",
+            size=12,
+        )
+        ax[1].legend()
+
+        # Finalize figure
+        title = f"{name.capitalize()} energy and reward distributions"
+        if step is not None:
+            title += f" (step {step})"
+        fig.suptitle(title, y=0.95)
+        plt.tight_layout(rect=[0, 0.02, 1, 0.98])
+
+        # store metrics
+        metrics = {
+            f"Mean {name} energy": mean_e,
+            f"Std {name} energy": std_e,
+            f"Mean {name} reward": mean_r,
+            f"Std {name} reward": std_r,
+            f"Mean {name} top k energy": mean_top_k_e,
+            f"Std {name} top k energy": std_top_k_e,
+            f"Mean {name} top k reward": mean_top_k_r,
+            f"Std {name} top k reward": std_top_k_r,
+            f"Best (min) {name} energy": best_e,
+            f"Best (max) {name} reward": best_r,
+        }
+        figs = [fig]
+        fig_names = [title]
+
+        if name.lower() == "train ground truth":
+            # train stats mode: the ground truth data has meen plotted
+            # and computed, let's do it again for the proxy data.
+            # This can be used to visualize potential distribution mismatch
+            # between the proxy and the ground truth data.
+            proxy_metrics, proxy_figs, proxy_fig_names = self.top_k_metrics_and_plots(
+                None,
+                top_k,
+                "train proxy",
+                energy=proxy,
+                reward=proxy_reward,
+                step=None,
+                **kwargs,
+            )
+            # aggregate metrics and figures
+            metrics.update(proxy_metrics)
+            figs += proxy_figs
+            fig_names += proxy_fig_names
+
+        return metrics, figs, fig_names
